Move RedisCountStorageTest into grakn-engine
