+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/manager/redisqueue/RedisTaskQueueConsumer.java
@@ -25,7 +25,7 @@ import static ai.grakn.engine.TaskStatus.STOPPED;
 import ai.grakn.engine.factory.EngineGraknTxFactory;
 import ai.grakn.engine.lock.LockProvider;
 import ai.grakn.engine.tasks.BackgroundTask;
-import ai.grakn.engine.tasks.connection.RedisCountStorage;
+import ai.grakn.engine.postprocessing.RedisCountStorage;
 import ai.grakn.engine.tasks.manager.TaskCheckpoint;
 import ai.grakn.engine.tasks.manager.TaskConfiguration;
 import ai.grakn.engine.tasks.manager.TaskState;
